coachloop/finetuning_data/embed_scripts/requirements.txt
fastapi==0.115.0 
uvicorn==0.30.6 
openai==1.51.0 
python-dotenv==1.0.1 
sounddevice==0.5.0 
numpy==1.26.4 
pydub==0.25.1 
wave==0.0.2
pip install fuzzywuzzy python-Levenshtein